half

+++ new.go
@@ -4,13 +4,45 @@
 package main
 
 import (
-	"bufio"
+	"encoding/gob"
 	"fmt"
+	"net"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"slices"
 	"strings"
 )
+
+type global_config struct {
+	Ipv4 string
+}
+
+type network_config struct {
+	Role            string
+	Wireless_status string
+	Ps              string
+	Ssid            string
+	Subnet          string
+	Name string
+}
+
+func get_interfaces() []string {
+	interfaces, err := net.Interfaces()
+	if err != nil {
+		fmt.Println("get interfaces failed", err)
+		return []string{}
+	}
+	interfac := []string{}
+	for _, iface := range interfaces {
+		if strings.Contains(iface.Name, "vir") || iface.Name == "lo" {
+			continue
+		}
+		interfac = append(interfac, iface.Name)
+	}
+	return interfac
+
+}
 
 const intent_dir string = "/etc/net/interfaces"
 const global_conf string = "/etc/net/global.conf"
@@ -40,15 +72,18 @@ func run(command string) {
 	}
 }
 
-func clean_slate() {
+func clean_slate(interfaces []string) {
 	fmt.Println("[*] Cleaning old configurations...")
 
-	files, _ := filepath.Glob(systemd_net_dir + "/*-enforced.network")
+	files, _ := filepath.Glob(systemd_net_dir + "/*.network")
 
 	for _, f := range files {
-		err := os.Remove(f)
-		if err != nil {
-			fmt.Println("failed to remove :", f, err)
+		baseName := filepath.Base(f)
+		if !slices.Contains(interfaces, strings.Split(baseName, ".")[0]) {
+			err := os.Remove(f)
+			if err != nil {
+				fmt.Println("failed to remove :", f, err)
+			}
 		}
 	}
 
@@ -78,53 +113,37 @@ func clean_slate() {
 
 }
 
-func parse_intent(path string) map[string]string {
-	config := make(map[string]string)
-	_, err := os.Stat(path)
-	if os.IsNotExist(err) {
-		return config
-	}
-	file, err := os.Open(path)
+func load_network(filepa string) network_config {
+	file, err := os.OpenFile(filepa, os.O_RDONLY, 0644)
 	if err != nil {
-		fmt.Println("cant open file:", err)
-		return config
+		fmt.Println("file error", err)
+		return network_config{}
 	}
 	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.Contains(line, "=") {
-			parts := strings.SplitN(line, "=", 2)
-			key := strings.TrimSpace(parts[0])
-			value := strings.TrimSpace(parts[1])
-
-			config[key] = value
-		}
-	}
-	return config
+	decoder := gob.NewDecoder(file)
+	var unpickled network_config
+	err = decoder.Decode(&unpickled)
+	return unpickled
 }
 
-func apply_global() bool {
-	fmt.Println("[*] Applying Global Settings...")
-	forward := "0"
+func apply_global() {
+	file, err := os.OpenFile(global_conf, os.O_RDONLY, 0644)
+	if err != nil {
+		fmt.Println("file error", err)
+		return
+	}
+	defer file.Close()
+	decoder := gob.NewDecoder(file)
+	var unpickled global_config
+	err = decoder.Decode(&unpickled)
+	var forward string
+	if unpickled.Ipv4 == "yes" {
+		forward = "1"
+	} else {
+		forward = "0"
 
-	file, err := os.Open(global_conf)
-	if err == nil {
-
-		defer file.Close()
-
-		scanner := bufio.NewScanner(file)
-		for scanner.Scan() {
-			line := scanner.Text()
-			if strings.Contains(line, "ip_forward=1") {
-				forward = "1"
-				break
-			}
-		}
 	}
 	run("sysctl -w net.ipv4.ip_forward=" + forward)
-	return forward == "1"
 
 }
 
@@ -133,8 +152,38 @@ func generate_subnet(index int) string {
 }
 
 func main() {
-	clean_slate()
-	ip_forwarding_enabled := apply_global()
+	interfaces := get_interfaces()
+	clean_slate(interfaces)
+	apply_global()
 	intent_files, _ := filepath.Glob(intent_dir + "/*.conf")
+
+	var final_intents []string
+	for _, file := range intent_files {
+		baseName := filepath.Base(file)
+		filet := strings.Split(baseName, ".")[0]
+		if slices.Contains(interfaces, filet) {
+			final_intents = append(final_intents, file)
+		}
+	}
+
+	for _ , final := range(final_intents){
+		current_intent := load_network(final)
+		if current_intent.Role == "uplink"{
+			metric := 200 
+            if strings.Contains(current_intent.Name , "et"){ 
+				metric = 100
+			}
+            if strings.Contains(current_intent.Name , "en"){ 
+				metric = 100
+			}
+			if strings.Contains(current_intent.Name , "wl"){ 
+				metric = 100
+			}
+			
+		}
+
+
+
+	}
 
 }
